requirements.txt
# ============================================
# Portfolio Manager — Dependencies
# ============================================

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
pandas-ta>=0.3.14b
scipy>=1.11.0

# Async & WebSocket
websockets>=12.0
aiohttp>=3.9.0

# Scoring & ML
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0

# Portfolio Optimization
PyPortfolioOpt>=1.5.5
riskfolio-lib>=6.0.0
empyrical>=0.5.5

# Database
SQLAlchemy>=2.0.0
redis>=5.0.0

# Scheduling
APScheduler>=3.10.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.1

# Telegram
python-telegram-bot>=20.6

# AI Reasoning
anthropic>=0.39.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0

# Logging
loguru>=0.7.0

# Config
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Finance data (fallback)
yfinance>=0.2.31
